+++ kaye_engine/abbr_collection/__init__.py
@@ -6,6 +6,7 @@ define various data structures supporting **abbreviation nodes**
 
 from .abbr_tags import AbbrTags
 from .abbr_wrap import AbbrWrap
+from .abbr_group import ABBRS_JSON_GROUP_KEY, AbbrGroupIndex
 from .abbr_data import AbbrData, get_abbr_data
 from .abbr_data_loader import populate_abbr_data_with_json_file
 from .abbr_meaning import AbbrMeaning
@@ -14,6 +15,8 @@ from .abbr_entry import AbbrEntry
 __all__ = (
     "AbbrTags",
     "AbbrWrap",
+    "ABBRS_JSON_GROUP_KEY",
+    "AbbrGroupIndex",
     "AbbrData",
     "get_abbr_data",
     "populate_abbr_data_with_json_file",