+++ tests/prompt/bp/prompt-bp-contains_test.py
@@ -368,23 +368,23 @@ class TestDynamicNodes:  #######################################################
     def test_obj1(_, dynamic_bp_testee1):
         bp = copy.deepcopy(dynamic_bp_testee1)
 
-        node = dynamic_bp_testee1.corpus["(Programming Languages Code)"]
+        node = dynamic_bp_testee1.corpus["(programming-language-codes)"]
 
         assert node in bp
 
     def test_hash1(_, dynamic_bp_testee1):
         bp = copy.deepcopy(dynamic_bp_testee1)
 
-        node = dynamic_bp_testee1.corpus["(Programming Languages Code)"]
+        node = dynamic_bp_testee1.corpus["(programming-language-codes)"]
 
         assert hash(node) in bp
 
     def test_name1(_, dynamic_bp_testee1):
         bp = copy.deepcopy(dynamic_bp_testee1)
 
-        assert "(Programming Languages Code)" in bp
+        assert "(programming-language-codes)" in bp
 
     def test_id1(_, dynamic_bp_testee1):
         bp = copy.deepcopy(dynamic_bp_testee1)
 
-        assert "(Programming Languages Code)" in bp
+        assert "(programming-language-codes)" in bp