+++ tests/prompt/conftest.py
@@ -14,9 +14,7 @@ from kaye_engine.prompt.prompt_corpus_node import PromptCorpusNode
 from kaye_engine.prompt import (
     TodayNode,
     AbbrNode,
-    UsableAbbrNode,
-    LanguageCodeNode,
-    PLCNode,
+    AbbrGroupNode,
 )
 
 
@@ -145,14 +143,15 @@ def corpus_testee3():
 def corpus_dynamic_testee(corpus_testee3):
     tree = deepcopy(corpus_testee3)
 
-    for node_type in (
-        TodayNode,
-        AbbrNode,
-        UsableAbbrNode,
-        LanguageCodeNode,
-        PLCNode,
-    ):
+    for node_type in (TodayNode, AbbrNode):
         node_type(tree)
+
+    for group_name in (
+        "usable-abbreviations",
+        "natural-language-codes",
+        "programming-language-codes",
+    ):
+        AbbrGroupNode(tree, group_name=group_name)
 
     return tree